requirements.txt
Flask==2.3.2
grpcio==1.54.2
pika==1.3.2
protobuf==4.23.2
reportlab==4.0.4
tk==0.1.0
requests==2.31.0
